+++ docs/src/conf.py
@@ -32,7 +32,13 @@ extensions = [
     'sphinx.ext.intersphinx', # generate links to the documentation of objects in external projects
     'sphinx.ext.mathjax', # Render math via JavaScript
     'sphinx.ext.napoleon', # Support for NumPy and Google style docstrings
+    'sphinxcontrib.bibtex', # Cite references using a bibtex file
 ]
+
+# The path to the bibtex file
+bibtex_bibfiles = ['../static/refs.bib']
+bibtex_default_style = 'unsrt'
+bibtex_reference_style = 'author_year'
 
 sphinx_gallery_conf = {
     "filename_pattern": "/*",