+++ tests/modules/test_temperatureprofile.py
@@ -21,10 +21,15 @@ class TestTemperaturProfile(ReferenceAtomGroups):
     def test_multiple(self, multiple_ags):
         """Test temperature."""
         temp = TemperaturePlanar(multiple_ags).run()
-        assert_allclose(temp.results.profile_mean[40], [223, 259], rtol=1e1)
+        assert_allclose(temp.results.profile[40], [223, 259], rtol=1e1)
 
     @pytest.mark.parametrize('dim', (0, 1, 2))
     def test_dens(self, ag, dim):
         """Test mean temperature."""
         dens = TemperaturePlanar(ag, dim=dim).run()
-        assert_allclose(dens.results.profile_mean.mean(), 291.6, rtol=1e1)
+        assert_allclose(dens.results.profile.mean(), 291.6, rtol=1e1)
+
+    def test_wrong_grouping(self, ag):
+        """Test a wrong grouping choice."""
+        with pytest.raises(ValueError, match="Invalid choice"):
+            TemperaturePlanar(ag, grouping="molecules").run()